- debugger: Use proper types in the API of the Disassembly dialog (int and uint64_t)
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@10540 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/include/cbdebugger_interfaces.h
@@ -62,10 +62,10 @@ class DLLIMPORT cbDisassemblyDlg
         virtual wxWindow* GetWindow() = 0;
 
         virtual void Clear(const cbStackFrame& frame) = 0;
-        virtual void AddAssemblerLine(size_t addr, const wxString& line) = 0;
-        virtual void AddSourceLine(size_t lineno, const wxString& line) = 0;
-        virtual bool SetActiveAddress(size_t addr) = 0;
-        virtual void CenterLine(size_t lineno) = 0;
+        virtual void AddAssemblerLine(uint64_t addr, const wxString& line) = 0;
+        virtual void AddSourceLine(int lineno, const wxString& line) = 0;
+        virtual bool SetActiveAddress(uint64_t addr) = 0;
+        virtual void CenterLine(int lineno) = 0;
         virtual void CenterCurrentLine() = 0;
         virtual bool HasActiveAddr() = 0;
         virtual void EnableWindow(bool enable) = 0;
